% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.crw.R
\name{aic.crw}
\alias{aic.crw}
\title{Calculates AIC for all objects of class crwFit listed as arguments}
\usage{
aic.crw(...)
}
\arguments{
\item{\dots}{a series of crwFit objects}
}
\value{
A table, sorted from lowest AIC value to highest.
}
\description{
AIC, delta AIC, and Akaike weights for all models listed as arguments.
}
\details{
The function can either be executed with a series of 'crwFit' objects (see
\code{\link{crwMLE}}) without the '.crwFit' suffix or the function can be
called without any arguments and it will search out all 'crwFit' objects in
the current workspace and produce the model selection table for all 'crwFit'
objects in the workspace. Caution should be used when executing the function
in this way. ALL 'crwFit' objects will be included whether ot not the same
locations are used!  For all of the models listed as arguments (or in the
workspace), AIC, delta AIC, and Akaike weights will be calculated.
}
\author{
Devin S. Johnson
}
