% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl-package.R
\docType{data}
\name{beardedSeals}
\alias{beardedSeals}
\title{Bearded Seal Location Data}
\format{
A data frame with 27,548 observations on 3 bearded seals in Alaska:

\describe{
\item{deployid}{Unique animal ID}
\item{ptt}{Hardware ID}
\item{instr}{Hardware type}
\item{date_time}{Time of location}
\item{type}{Location type}
\item{quality}{Argos location quality}
\item{latitude}{Observed latitude}
\item{longitude}{Observed longitude}
\item{error_radius}{Argos error radius}
\item{error_semimajor_axis}{Argos error ellipse major axis length}
\item{error_semiminor_axis}{Argos error ellipse minor axis length}
\item{error_ellipse_orientation}{Argos error ellispse degree orientation}
}
}
\source{
Marine Mammal Laboratory, Alaska
Fisheries Science Center, National Marine Fisheries Service, NOAA 7600 Sand
Point Way NE Seattle, WA 98115
}
\description{
Bearded Seal Location Data
}
\keyword{datasets}
