% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_shiny_apps.R
\name{check_csv}
\alias{check_csv}
\title{Start a shiny app to check data stored in a .csv file for model fitting with \code{crwMLE} function.}
\usage{
check_csv()
}
\description{
Start a shiny app to check data stored in a .csv file for model fitting with \code{crwMLE} function.
}

