% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crw_coerce.R
\name{coerce-crw-methods}
\alias{coerce-crw-methods}
\alias{coerce-crw}
\alias{crw_as_tibble}
\alias{crw_as_tibble.crwIS}
\alias{crw_as_tibble.crwPredict}
\alias{coerce-crw}
\alias{crw_as_sf_point}
\alias{crw_as_sf_point.crwIS}
\alias{crw_as_sf_point.crwPredict}
\alias{coerce-crw}
\alias{crw_as_sf_line}
\alias{crw_as_sf_line.crwIS}
\alias{crw_as_sf_line.crwPredict}
\title{Coerce crawl objects (crwIS and crwPredict) to various objects}
\usage{
crw_as_tibble(crw_object)

\method{crw_as_tibble}{crwIS}(crw_object)

\method{crw_as_tibble}{crwPredict}(crw_object)

crw_as_sf_point(crw_object, epsg)

\method{crw_as_sf_point}{crwIS}(crw_object, epsg)

\method{crw_as_sf_point}{crwPredict}(crw_object, epsg)

crw_as_sf_line(crw_object, epsg)

\method{crw_as_sf_line}{crwIS}(crw_object, epsg)

\method{crw_as_sf_line}{crwPredict}(crw_object, epsg)
}
\arguments{
\item{crw_object}{an object of class \code{"crwIS"} or \code{"crwPredict"},}

\item{crw_object}{an object of class \code{"crwIS"} or \code{"crwPredict"},}

\item{crw_object}{an object of class \code{"crwIS"} or \code{"crwPredict"},}
}
\value{
\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}

\code{NULL}
}
\description{
Coerce crawl objects (crwIS and crwPredict) to various objects

Coerce to tibble

Coerce to sf point object

Coerce to sf line object
}
\author{
Josh M. London
}
