% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_path.R
\name{crw_alpha}
\alias{crw_alpha}
\title{Extract alpha values from \code{crwPredict} or \code{crwIS} objects}
\usage{
crw_alpha(crw_object)
}
\arguments{
\item{crw_object}{Can be one of the following classes: 
(1) 'crwIS' object from the \code{crwPostIS} function
(2) 'crwPredict' object from the \code{crwPredict} function}
}
\value{
a list of 2 (matrix of extracted alpha values, times)
}
\description{
extracts the needed alpha values for \code{fix_path}
}
