% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crw_coerce_sf.R
\name{crw_as_sf}
\alias{crw_as_sf}
\alias{crw_as_sf.crwIS}
\alias{crw_as_sf.crwPredict}
\title{Coerce to sf/sfc object}
\usage{
crw_as_sf(crw_object, epsg, ftype, group)

\method{crw_as_sf}{crwIS}(crw_object, epsg, ftype, group)

\method{crw_as_sf}{crwPredict}(crw_object, epsg, ftype, group)
}
\arguments{
\item{crw_object}{an object of class \code{"crwIS"} or \code{"crwPredict"}}

\item{epsg}{integer epsg code specifying the coordinate projection}

\item{ftype}{character of either "POINT" or "LINESTRING" specifying the feature type}

\item{group}{character specifying the column to group by for mulitple LINESTRING features}
}
\description{
Coerce to sf/sfc object
}
\section{Methods (by class)}{
\itemize{
\item \code{crwIS}: coerce crwIS object to sf (POINT geometry) or sfc_LINESTRING

\item \code{crwPredict}: coerce crwPredict object to sf (POINT geometry) or sfc_LINESTRING
}}

