% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crw_coerce_tibble.R
\name{crw_as_tibble}
\alias{crw_as_tibble}
\alias{crw_as_tibble.crwIS}
\alias{crw_as_tibble.crwPredict}
\alias{crw_as_tibble.tbl}
\title{Coerce crawl objects (crwIS and crwPredict) to tibbles}
\usage{
crw_as_tibble(crw_object, ...)

\method{crw_as_tibble}{crwIS}(crw_object, ...)

\method{crw_as_tibble}{crwPredict}(crw_object, ...)

\method{crw_as_tibble}{tbl}(crw_object, ...)
}
\arguments{
\item{crw_object}{an object of class \code{"crwIS"} or \code{"crwPredict"}}

\item{...}{Additional arguments that are ignored}
}
\description{
Coerce crawl objects (crwIS and crwPredict) to tibbles
}
\section{Methods (by class)}{
\itemize{
\item \code{crwIS}: coerce crwIS object to tibble

\item \code{crwPredict}: coerce crwPredict object to tibble

\item \code{tbl}: 
}}

\author{
Josh M. London
}
