% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crw_broom.R
\name{crw_broom}
\alias{crw_broom}
\title{tidy-like method for crwFit object}
\usage{
crw_broom(fit)
}
\arguments{
\item{fit}{\code{crwFit} object from \code{crawl::crwMLE}}
}
\description{
this function mimics the approach taken by \code{broom::tidy}
to present model output parameters in a tidy, data frame structure.
}
