% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwMLE.R
\name{crwMLE}
\alias{crwMLE}
\title{Fit Continuous-Time Correlated Random Walk Models to Animal Telemetry Data}
\usage{
crwMLE(mov.model = ~1, err.model = NULL, activity = NULL,
  drift = FALSE, data, coord = c("x", "y"), Time.name = "time",
  time.scale = NULL, theta, fixPar, method = "Nelder-Mead",
  control = NULL, constr = list(lower = -Inf, upper = Inf),
  prior = NULL, need.hess = TRUE, initialSANN = list(maxit = 200),
  attempts = 1, ...)
}
\arguments{
\item{mov.model}{formula object specifying the time indexed covariates for
movement parameters.}

\item{err.model}{A 2-element list of formula objects specifying the time
indexed covariates for location error parameters.}

\item{activity}{formula object giving the covariate for the activity (i.e., stopped or fully moving)
portion of the model.}

\item{drift}{logical indicating whether or not to include a random
drift component. For most data this is usually not necessary. See \code{\link{northernFurSeal}} for an example
using a drift model.}

\item{data}{data.frame object containg telemetry and covariate data. A 
'sf' object from the 'sf' package with a geometry column of type \code{sfc_POINT}
is the preferred format for these data. 'SpatialPointsDataFrame' object 
from the package 'sp' is still accepted. `spacetime' and 'trip' objects were 
previously accepted but are no longer valid. Values for coords will be taken from 
the spatial data set and ignored in the arguments. Spatial data must have a
valid proj4string or epsg and must NOT be in longlat. While data provided as
a data.frame are supported, they will not be compatible with additional 
fucntions e.g. \code{fix_path}.}

\item{coord}{A 2-vector of character values giving the names of the "X" and
"Y" coordinates in \code{data}.}

\item{Time.name}{character indicating name of the location time column. It is
strongly preferred that this column be of type POSIXct and in UTC.}

\item{time.scale}{character. Scale for conversion of POSIX time to numeric 
for modeling. Defaults to "hours" and most users will not need to change this.}

\item{theta}{starting values for parameter optimization.}

\item{fixPar}{Values of parameters which are held fixed to the given value.}

\item{method}{Optimization method that is passed to \code{\link{optim}}.}

\item{control}{Control list which is passed to \code{\link{optim}}.}

\item{constr}{Named list with elements \code{lower} and \code{upper} that
are vectors the same length as theta giving the box constraints for the
parameters}

\item{prior}{A function returning the log-density function of the parameter
prior distribution. THIS MUST BE A FUNCTION OF ONLY THE FREE PARAMETERS. Any 
fixed parameters should not be included.}

\item{need.hess}{A logical value which decides whether or not to evaluate
the Hessian for parameter standard errors}

\item{initialSANN}{Control list for \code{\link{optim}} when simulated
annealing is used for obtaining start values. See details}

\item{attempts}{The number of times likelihood optimization will be
attempted}

\item{...}{Additional arguments that are ignored.}
}
\value{
A list with the following elements:

\item{par}{Parameter maximum likelihood estimates (including fixed
parameters)}

\item{estPar}{MLE without fixed parameters}

\item{se}{Standard error of MLE}

\item{ci}{95\% confidance intervals for parameters}

\item{Cmat}{Parameter covariance matrix}

\item{loglik}{Maximized log-likelihood value}

\item{aic}{Model AIC value}

\item{coord}{Coordinate names provided for fitting}

\item{fixPar}{Fixed parameter values provided}

\item{convergence}{Indicator of convergence (0 = converged)}

\item{message}{Meesages given by \code{optim} during parameter optimization}

\item{activity}{Model provided for stopping variable}

\item{drift}{Logical value indicating random drift model}

\item{mov.model}{Model description for movement component}

\item{err.model}{Model description for location error component}

\item{n.par}{number of parameters}

\item{nms}{parameter names}

\item{n.mov}{number of movement parameters}

\item{n.errX}{number or location error parameters for "longitude" error
model}

\item{n.errY}{number or location error parameters for "latitude" error
model}

\item{stop.mf}{covariate for stop indication in stopping models}

\item{polar.coord}{Logical indicating coordinates are polar latitude and
longitude}

\item{init}{Initial values for parameter optimization}

\item{data}{Original data.frame used to fit the model}

\item{lower}{The lower parameter bounds}

\item{upper}{The upper parameter bounds}

\item{need.hess}{Logical value}

\item{runTime}{Time used to fit model}
}
\description{
The function uses the Kalman filter to estimate movement paramters in a
state-space version of the continuous-time movement model. Separate models
are specified for movement portion and the location error portion. Each
model can depend on time indexed covariates. A \dQuote{haul out} model where
movement is allowed to completely stop, as well as, a random drift model can
be fit with this function.
}
\author{
Devin S. Johnson, Josh M. London
}
