% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwSimulator.R
\name{crwSimulator}
\alias{crwSimulator}
\title{Construct a posterior simulation object for the CTCRW state vectors}
\usage{
crwSimulator(
  object.crwFit,
  predTime = NULL,
  method = "IS",
  parIS = 1000,
  df = Inf,
  grid.eps = 1,
  crit = 2.5,
  scale = 1,
  quad.ask = TRUE,
  force.quad
)
}
\arguments{
\item{object.crwFit}{A model object from \code{\link{crwMLE}}.}

\item{predTime}{vector of additional prediction times.}

\item{method}{Method for obtaining weights for movement parameter samples}

\item{parIS}{Size of the parameter importance sample}

\item{df}{Degrees of freedom for the t approximation to the parameter
posterior}

\item{grid.eps}{Grid size for \code{method="quadrature"}}

\item{crit}{Criterion for deciding "significance" of quadrature points
(difference in log-likelihood)}

\item{scale}{Scale multiplier for the covariance matrix of the t
approximation}

\item{quad.ask}{Logical, for method='quadrature'. Whether or not the sampler should ask if quadrature sampling should take place.
It is used to stop the sampling if the number of likelihood evaluations would be extreme.}

\item{force.quad}{A logical indicating whether or not to force the execution 
of the quadrature method for large parameter vectors.}
}
\value{
List with the following elements:

\item{x}{Longitude coordinate with NA at prediction times}

\item{y}{Similar to above for latitude}

\item{locType}{Indicates prediction types with a "p" or observation times
with an "o"} \item{P1.y}{Initial state covariance for latitude}

\item{P1.x}{Initial state covariance for longitude}

\item{a1.y}{Initial latitude state}

\item{a1.x}{Initial longitude state}

\item{n.errX}{number of longitude error model parameters}

\item{n.errY}{number of latitude error model parameters}

\item{delta}{vector of time differences}

\item{driftMod}{Logical. indicates random drift model}

\item{stopMod}{Logical. Indicated stop model fitted}

\item{stop.mf}{stop model design matrix}

\item{err.mfX}{Longitude error model design matrix}

\item{err.mfY}{Latitude error model design matrix}

\item{mov.mf}{Movement model design matrix}

\item{fixPar}{Fixed values for parameters in model fitting}

\item{Cmat}{Covaraince matrix for parameter sampling distribution}

\item{Lmat}{Cholesky decomposition of Cmat}

\item{par}{fitted parameter values}

\item{N}{Total number of locations}

\item{loglik}{log likelihood of the fitted model}

\item{Time}{vector of observation times}

\item{coord}{names of coordinate vectors in original data}

\item{Time.name}{Name of the observation times vector in the original data}

\item{thetaSampList}{A list containing a data frame of parameter vectors and
their associated probabilities for a resample}
}
\description{
The \code{crwSimulator} function uses a fitted model object from
\code{crwMLE} and a set of prediction times to construct a list from which
\code{\link{crwPostIS}} will draw a sample from either the posterior
distribution of the state vectors conditional on fitted parameters or a full
posterior draw from an importance sample of the parameters.
}
\details{
The crwSimulator function produces a list and preprocesses the necessary
components for repeated track simulation from a fitted CTCRW model from
\code{\link{crwMLE}}. The \code{method} argument can be one of \code{"IS"}
or \code{"quadrature"}. If method="IS" is chosen standard importance
sampling will be used to calculate the appropriate weights via t proposal
with df degrees of freedom.  If df=Inf (default) then a multivariate normal
distribution is used to approximate the parameter posterior. If
\code{method="quadrature"}, then a regular grid over the posterior is used
to calculate the weights. The argument \code{grid.eps} controls the
quadrature grid. The arguments are approximately the upper and lower limit
in terms of standard deviations of the posterior.  The default is
\code{grid.eps}, in units of 1sd. If \code{object.crwFit} was fitted with
\code{crwArgoFilter}, then the returned list will also include \code{p.out},
which is the approximate probability that the observation is an outlier.
}
\seealso{
See \code{demo(northernFurSealDemo)} for example.
}
\author{
Devin S. Johnson
}
