% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crwUseGrid.R
\name{crwUseGrid}
\alias{crwUseGrid}
\title{Compute a spatial use grid from a crawl prediction}
\usage{
crwUseGrid(object, grid, subset = TRUE)
}
\arguments{
\item{object}{A 'SpatialPoints' object created with the \code{sp} package.  T}

\item{grid}{A \code{GridTopology} object from the 'sp' package}

\item{subset}{An indicator of which times should be used for calculation of
the use grid. Can be a logical vector or a vector of integers, such as from
a call to \code{which}}
}
\value{
A \code{SpatialGridDataFrame} with data column 'use' which gives the
count of locations within the grid cell
}
\description{
This function take a SpatialPoints object and a spatial GridTopology
object from the 'sp' package and outputs the number of point locations
in each grid cell
}
\author{
Devin S. Johnson <devin.johnson@noaa.gov>
}

