% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_timescale.R
\name{detect_timescale}
\alias{detect_timescale}
\title{Detect appropriate time scale for movement analysis}
\usage{
detect_timescale(time_vector)
}
\arguments{
\item{time_vector}{a vector of class POSIXct}
}
\value{
character of either "seconds","minutes","hours","days","weeks"
}
\description{
This function examines the time vector and evaluates the median time
interval. With this, we determine what the best time scale for the 
movement model is likely to be.
}
