% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_paths.R
\name{fix_path}
\alias{fix_path}
\title{Project path away from restricted areas}
\usage{
fix_path(xy, res_raster, trans)
}
\arguments{
\item{xy}{Coordinate locations for the path. Can be one of the following classes: 
(1) a two column matrix, 
(2) 'SpatialPoints' or 'SpatialPointsDataFrame' object from the sp package,
(3) 'crwPredict' object from the \code{crwPredict} function
(4) 'crwIS' object from the \code{crwPostIS} function}

\item{res_raster}{An indicator raster object with cells = 1 if it is 'off-limits'
and 0 elsewise.}

\item{trans}{A transition matrix object from the gdistance package.}
}
\value{
Either matrix or 'SpatialPoints' object with path projected around
restricted areas
}
\description{
Corrects a path so that it does not travel through a restricted area.
}

