% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_path.R
\name{fix_segments}
\alias{fix_segments}
\title{Identify and re-route segments of a path that cross a restricted area}
\usage{
fix_segments(
  crw_sf,
  vector_mask,
  barrier_buffer = 50,
  crwFit,
  alpha,
  crwIS = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{crw_sf}{an `sf` object derived from a \code{crwIS} or \code{crwPredict}}

\item{vector_mask}{an 'sf' polygon object that defines the restricted area}

\item{crwFit}{crwFit object that was used to generate the \code{crwIS} or \code{crwPredict}}
}
\value{
a tibble with each record identifying the segments and pertinant values
}
\description{
This function takes an `sf` object derived from a \code{crwIS} or 
\code{crwPredict} and an `sf` polygon object that defines the restricted
area. Each segment of the path that crosses the restricted area is identified
using the \code{get_mask_segments} function. Each segment begins and ends with
a coordinate that is outside the restricted area. The path is re-routed using
a mix of brownian movement and correlated parameters from the fitted movement
model (\code{crwFit}). A regular distribution of waypoints are established along
the restricted area boundary to guide the movement path around the area.
}
