% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethod.R
\name{flatten}
\alias{flatten}
\title{'Flattening' a list-form crwPredict object into a data.frame}
\usage{
flatten(predObj)
}
\arguments{
\item{predObj}{A crwPredict object}
}
\value{
a \code{\link{data.frame}} version of a crwPredict list with columns
for the state standard errors
}
\description{
\dQuote{Flattens} a list form \code{\link{crwPredict}} object into a flat
data.frame.
}
\seealso{
\code{\link{northernFurSeal}} for use example
}
\author{
Devin S. Johnson
}
