% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_path.R
\name{get_mask_segments}
\alias{get_mask_segments}
\title{Identify segments of a path that cross through a land barrier}
\usage{
get_mask_segments(crw_sf, vector_mask, alpha)
}
\arguments{
\item{crw_sf}{A \code{crwIS} or \code{crwPredict} object that has been converted
to an \code{sf} 'POINT' object via the \code{crw_as_sf} function.}

\item{vector_mask}{A \code{sf} object from sf package that indicates
restricted areas as a polygon feature.}

\item{alpha}{alpha values extracted from a \code{crwPredict} or \code{crwIS}
object via the \code{crw_alpha} function.}
}
\value{
A data.frame with each row associated with each section of the path
  that crosses a restricted area. The columns provide the start and end row
  indices of \code{xy} where the section occurs and the previous and post
  locations that are in unrestricted space.
}
\description{
This function is used to identify sections of a path that cross
  through a land barrier. the \code{crwMLE} function in \code{crawl} cannot
  actively steer paths away from land. So, this function will identify path
  segments from the unrestrained path.
}
\author{
Josh M. London (josh.london@noaa.gov)
}
