% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logSpeed}
\alias{logSpeed}
\title{calculate estimate of log speed from a crwPredict object}
\usage{
logSpeed(alpha.hat, V.hat)
}
\arguments{
\item{alpha.hat}{State estimates from \code{\link{crwPredict}}}

\item{V.hat}{An array of covariance matrices for alpha.hat.}
}
\value{
data.frame with colums:

\item{ln.speed}{log speed estimate (in meters/time unit)}
\item{var.ln.speed}{log speed estimated standard error.}
}
\description{
Calculates log speed estimates and standard errors based on the delta
method. Typically this is used within the function \code{\link{crwPredict}}.
}
\author{
Devin S. Johnson
}

