% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crw-methods.R
\name{[.crwIS}
\alias{[.crwIS}
\title{Generic subset/bracket method for crwIS classes}
\usage{
\method{[}{crwIS}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{...}{other arguments}
}
\description{
Generic subset/bracket method for crwIS classes
}
