% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_crwFit.R
\name{tidy_crwFit}
\alias{tidy_crwFit}
\title{tidy-like method for crwFit object}
\usage{
tidy_crwFit(fit)
}
\arguments{
\item{fit}{\code{crwFit} object from \code{crawl::crwMLE}}
}
\description{
this function mimics the approach taken by \code{broom::tidy}
to present model output parameters in a tidy, data frame structure.
}
