\name{crawl-package}
\docType{package}

\alias{crawl-package}
\alias{crawl}

% -------------------------------------------------------------------------

\title{Fit Continuous-Time Correlated Random Walk models to animal movement data}


\description{ The (C)orrelated (RA)ndom (W)alk (L)ibrary (I know it is not 
  an R library, but, "crawp" did not sound as good) of R functions was designed
  for fitting continuous-time correlated random walk (CTCRW) models with
  time indexed covariates. The model is fit using the Kalman-Filter on a
  state space version of the continuous-time staochistic movement
  process. }


\details{

  \tabular{ll}{
    Package: \tab crawl\cr
    Type: \tab Package\cr
    Version: \tab 1.3\cr
    Date: \tab 2011-05-26\cr
    License: \tab Unlimited \cr
    LazyLoad: \tab yes\cr
  }

}


\references{ Johnson, D., J. London, M. -A. Lea, and J. Durban (2008)
Continuous-time correlated random walk model for animal telemetry
data. Ecology 89(5) 1208-1215. }

% -------------------------------------------------------------------------

\author{
  Devin S. Johnson

  Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}
