% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crawl-package.R
\docType{package}
\name{crawl-package}
\alias{crawl}
\alias{crawl-package}
\title{Fit Continuous-Time Correlated Random Walk Models to Animal Movement Data}
\description{
The (C)orrelated (RA)ndom (W)alk (L)ibrary (I know it is not an R library,
but, "crawp" did not sound as good) of R functions was designed for fitting
continuous-time correlated random walk (CTCRW) models with time indexed
covariates. The model is fit using the Kalman-Filter on a state space
version of the continuous-time staochistic movement process.
}
\details{
\tabular{ll}{ Package: \tab crawl\cr Type: \tab Package\cr Version: \tab
1.5\cr Date: \tab Aug 19, 2015\cr License: \tab Unlimited \cr LazyLoad: \tab
yes\cr }
}
\author{
Devin S. Johnson

Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}
\references{
Johnson, D., J. London, M. -A. Lea, and J. Durban (2008)
Continuous-time correlated random walk model for animal telemetry data.
Ecology 89(5) 1208-1215.
}

