% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl-package.R
\docType{package}
\name{crawl-package}
\alias{crawl-package}
\alias{crawl}
\title{Fit Continuous-Time Correlated Random Walk Models to Animal Movement Data}
\description{
The Correlated RAndom Walk Library (I know it is not an R library,
but, "crawp" did not sound as good) of R functions was designed for fitting
continuous-time correlated random walk (CTCRW) models with time indexed
covariates. The model is fit using the Kalman-Filter on a state space
version of the continuous-time staochistic movement process.

\tabular{ll}{ 
Package: \tab crawl\cr 
Type: \tab Package\cr 
Version: \tab 2.2.2\cr 
Date: \tab September 17, 2018\cr 
License: \tab CC0 \cr 
LazyLoad: \tab yes\cr 
}
}
\note{
This software package is developed and maintained by scientists at the NOAA Fisheries Alaska 
Fisheries Science Center and should be considered a fundamental research communication. 
The reccomendations and conclusions presented here are those of 
the authors and this software should not be construed as official communication by NMFS, NOAA, 
or the U.S. Dept. of Commerce. In addition, reference to trade names does not imply endorsement by the 
National Marine Fisheries Service, NOAA. While the best efforts have been made to insure the 
highest quality, tools such as this are under constant development and are subject to change.
}
\references{
Johnson, D., J. London, M. -A. Lea, and J. Durban (2008)
Continuous-time correlated random walk model for animal telemetry data.
Ecology 89(5) 1208-1215.
}
\author{
Devin S. Johnson

Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}
