% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crwSamplePar.R
\name{crwSamplePar}
\alias{crwSamplePar}
\title{Create a weighted importance sample for posterior predictive track
simulation.}
\usage{
crwSamplePar(
  object.sim,
  method = "IS",
  size = 1000,
  df = Inf,
  grid.eps = 1,
  crit = 2.5,
  scale = 1,
  quad.ask = T,
  force.quad
)
}
\arguments{
\item{object.sim}{A simulation object from \code{\link{crwSimulator}}.}

\item{method}{Method for obtaining weights for movement parameter samples}

\item{size}{Size of the parameter importance sample}

\item{df}{Degrees of freedom for the t approximation to the parameter
posterior}

\item{grid.eps}{Grid size for \code{method="quadrature"}}

\item{crit}{Criterion for deciding "significance" of quadrature points
(difference in log-likelihood)}

\item{scale}{Scale multiplier for the covariance matrix of the t
approximation}

\item{quad.ask}{Logical, for method='quadrature'. Whether or not the sampler should ask if quadrature sampling should take place.
It is used to stop the sampling if the number of likelihood evaluations would be extreme.}

\item{force.quad}{A logical indicating whether or not to force the execution 
of the quadrature method for large parameter vectors.}
}
\value{
List with the following elements:

\item{x}{Longitude coordinate with NA at prediction times}

\item{y}{Similar to above for latitude}

\item{locType}{Indicates prediction types with a "p" or observation times
with an "o"} \item{P1.y}{Initial state covariance for latitude}

\item{P1.x}{Initial state covariance for longitude}

\item{a1.y}{Initial latitude state}

\item{a1.x}{Initial longitude state}

\item{n.errX}{number of longitude error model parameters}

\item{n.errY}{number of latitude error model parameters}

\item{delta}{vector of time differences}

\item{driftMod}{Logical. indicates random drift model}

\item{stopMod}{Logical. Indicated stop model fitted}

\item{stop.mf}{stop model design matrix}

\item{err.mfX}{Longitude error model design matrix}

\item{err.mfY}{Latitude error model design matrix}

\item{mov.mf}{Movement model design matrix}

\item{fixPar}{Fixed values for parameters in model fitting}

\item{Cmat}{Covaraince matrix for parameter sampling distribution}

\item{Lmat}{Cholesky decomposition of Cmat}

\item{par}{fitted parameter values}

\item{N}{Total number of locations}

\item{loglik}{log likelihood of the fitted model}

\item{Time}{vector of observation times}

\item{coord}{names of coordinate vectors in original data}

\item{Time.name}{Name of the observation times vector in the original data}

\item{thetaSampList}{A list containing a data frame of parameter vectors and
their associated probabilities for a resample}
}
\description{
The \code{crwSamplePar} function uses a fitted model object from
\code{crwMLE} and a set of prediction times to construct a list from which
\code{\link{crwPostIS}} will draw a sample from either the posterior
distribution of the state vectors conditional on fitted parameters or a full
posterior draw from an importance sample of the parameters.
}
\details{
The crwSamplePar function uses the information in a
\code{\link{crwSimulator}} object to create a set of weights for importance
sample-resampling of parameters in a full posterior sample of parameters and
locations using \code{\link{crwPostIS}}. This function is usually called
from \code{\link{crwPostIS}}. The average user should have no need to call
this function directly.
}
\seealso{
See \code{demo(northernFurSealDemo)} for example.
}
\author{
Devin S. Johnson
}
