% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_path.R
\name{fix_path}
\alias{fix_path}
\title{Project path away from restricted areas}
\usage{
fix_path(crw_object, vector_mask, crwFit)
}
\arguments{
\item{crw_object}{Coordinate locations for the path. Can be one of the following classes: 
(1) 'crwIS' object from the \code{crwPostIS} function}

\item{vector_mask}{an 'sf' polygon object that defines the restricted area}

\item{crwFit}{crwFit object that was used to generate the crw_object}
}
\value{
a new crw_object (of type crwIS)
}
\description{
Corrects a path so that it does not travel through a restricted area.
}
