% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_path.R
\name{fix_segments}
\alias{fix_segments}
\title{Identify segments of a path that cross a restricted area}
\usage{
fix_segments(crw_object, vector_mask, crwFit)
}
\arguments{
\item{crw_object}{Coordinate locations for the path. Can be one of the following classes: 
(1) 'crwIS' object from the \code{crwPostIS} function}

\item{vector_mask}{an 'sf' polygon object that defines the restricted area}

\item{crwFit}{crwFit object that was used to generate the crw_object}
}
\value{
a tibble with each record identifying the segments and pertinant values
}
\description{
This function takes a crw_object (crwIS only for now) and an 'sf'
polygon object that defines the restricted area and identifies each segment
of the path that crosses the restricted area. Each segment begins and ends with
a coordinate that is outside the restricted area.
}
