% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl-package.R
\docType{data}
\name{northernFurSeal}
\alias{northernFurSeal}
\title{Northern fur seal pup relocation data set used in Johnson et al. (2008)}
\format{
A data frame with 795 observations on the following 4 variables:

\describe{ \item{GMT}{A POSIX time vector}

\item{loc_class}{a factor with levels \code{3} \code{2}
\code{1} \code{0} \code{A}.}

\item{lat}{a numeric vector. Latitude for the locations}

\item{long}{a numeric vector. Longitude for the locations}

}
}
\source{
Marine Mammal Laboratory, Alaska
Fisheries Science Center, National Marine Fisheries Service, NOAA 7600 Sand
Point Way NE Seattle, WA 98115
}
\description{
Northern fur seal pup relocation data set used in Johnson et al. (2008)
}
\references{
Johnson, D., J. London, M. -A. Lea, and J. Durban (2008) Continuous-time
random walk model for animal telemetry data. Ecology 89:1208-1215.
}
\keyword{datasets}
