% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crw-methods.R
\name{[.crwIS}
\alias{[.crwIS}
\title{Generic methods for crwIS and crwPredict classes}
\usage{
\method{[}{crwIS}(x, i, j, drop_obs = FALSE, ..., drop = TRUE)
}
\arguments{
\item{drop_obs}{boolean whether to drop the "o" locType records}

\item{...}{other arguments}
}
\description{
Generic methods for crwIS and crwPredict classes
}
